refactored code in HeaderComponent, seperated out Footer component and fixed header menu issue. Deleted reference components and deleted ref code in Contact component
