+++ README.md
@@ -12,8 +12,6 @@ yarn start/npm start
 ```
 
 
-This project was bootstrapped with [Create React App](https://github.com/facebookincubator/create-react-app).
-
 Below you will find some information on how to perform common tasks.<br>
 You can find the most recent version of this guide [here](https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/template/README.md).
