+++ src/components/About.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { Grid, Header } from "semantic-ui-react";
+import { Grid, Header, Divider } from "semantic-ui-react";
 import "semantic-ui-css/semantic.min.css";
 
 export default class About extends Component {
@@ -9,9 +9,10 @@ export default class About extends Component {
         <Grid stackable container>
           <Grid.Row>
             <Grid.Column>
-              <Header as="h1" textAlign="center" style={{ fontSize: "2.3em" }}>
+              <Header as='h2' textAlign="left">
                 About Us
               </Header>
+              <Divider />
               <p style={{ fontSize: "1.33em" }}>
                 We are a group of people, who are working 100% to bring lawyers
                 and inventors on one platform to make their work easier for both
