+++ src/components/App.js
@@ -1,28 +1,31 @@
-import React, { Component } from "react";
+import React, { Component, Fragment } from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
 import HomepageLayout from "./HomepageLayout";
-import ResponsiveContainer from "./HeaderComponent";
 import About from "./About";
 import Contact from "./Contact";
 import FAQ from "./Faq";
 import TermsAndConditions from "./TermsAndConditions";
 import PrivacyPolicy from "./PrivacyPolicy";
+import Footer from "./Footer";
+import Header from "./Header";
 
 class App extends Component {
   render() {
     return (
       <Router>
-        <ResponsiveContainer>
-          <div>
-            <Route exact path="/" component={HomepageLayout} />
-            <Route path="/about" component={About} />
-            <Route path="/contact" component={Contact} />
-            <Route path="/faq" component={FAQ} />
-            <Route path="/tnc" component={TermsAndConditions} />
-            <Route path="/privacy" component={PrivacyPolicy} />
-          </div>
-        </ResponsiveContainer>
+        <Fragment>
+          <Header />
+
+          <Route exact path="/" component={HomepageLayout} />
+          <Route path="/about" component={About} />
+          <Route path="/contact" component={Contact} />
+          <Route path="/faq" component={FAQ} />
+          <Route path="/tnc" component={TermsAndConditions} />
+          <Route path="/privacy" component={PrivacyPolicy} />
+
+          <Footer />
+        </Fragment>
       </Router>
     );
   }
